+++ app/controllers/answers_controller.rb
@@ -6,8 +6,14 @@ class AnswersController < ApplicationController
     @step = Step.find(step_id)
 
     @answer = AnswerFactory.new(step: @step).call
-    @answer.assign_attributes(answer_params)
     @answer.step = @step
+
+    case @step.contentful_type
+    when "checkboxes"
+      @answer.assign_attributes(checkbox_params)
+    else
+      @answer.assign_attributes(answer_params)
+    end
 
     if @answer.valid?
       @answer.save
@@ -17,7 +23,22 @@ class AnswersController < ApplicationController
         redirect_to journey_path(@journey)
       end
     else
-      render "steps/new.#{@step.contentful_type}"
+      render "steps/#{@step.contentful_type}", locals: {layout: "steps/new_form_wrapper"}
+    end
+  end
+
+  def update
+    @journey = Journey.find(journey_id)
+    @step = Step.find(step_id)
+    @answer = @step.answer
+
+    @answer.assign_attributes(answer_params)
+
+    if @answer.valid?
+      @answer.save
+      redirect_to journey_path(@journey)
+    else
+      render "steps/#{@step.contentful_type}", locals: {layout: "steps/edit_form_wrapper"}
     end
   end
 
@@ -33,5 +54,9 @@ class AnswersController < ApplicationController
 
   def answer_params
     params.require(:answer).permit(:response)
+  end
+
+  def checkbox_params
+    params.require(:answer).permit(response: [])
   end
 end