+++ app/controllers/journeys_controller.rb
@@ -8,8 +8,9 @@ class JourneysController < ApplicationController
 
   def show
     @journey = Journey.includes(
-      steps: [:radio_answer, :short_text_answer]
+      steps: [:radio_answer, :short_text_answer, :long_text_answer]
     ).find(journey_id)
+    @steps = @journey.steps.map { |step| StepPresenter.new(step) }
   end
 
   private