+++ app/controllers/steps_controller.rb
@@ -31,7 +31,15 @@ class StepsController < ApplicationController
     @step = Step.find(params[:id])
     @answer = AnswerFactory.new(step: @step).call
 
-    render "new.#{@step.contentful_type}"
+    render @step.contentful_type, locals: {layout: "steps/new_form_wrapper"}
+  end
+
+  def edit
+    @journey = Journey.find(journey_id)
+    @step = Step.find(params[:id])
+    @answer = @step.answer
+
+    render "steps/#{@step.contentful_type}", locals: {layout: "steps/edit_form_wrapper"}
   end
 
   private