+++ app/services/answer_factory.rb
@@ -10,6 +10,8 @@ class AnswerFactory
     when "radios" then RadioAnswer.new
     when "short_text" then ShortTextAnswer.new
     when "long_text" then LongTextAnswer.new
+    when "single_date" then SingleDateAnswer.new
+    when "checkboxes" then CheckboxAnswers.new
     end
   end
 end