+++ config/initializers/sidekiq.rb
@@ -1,0 +1,13 @@
+Sidekiq.configure_server do |config|
+  config.redis = {url: ENV["REDIS_SIDEKIQ_URL"]}
+
+  # Sidekiq Cron
+  schedule_file = "config/schedule.yml"
+  if File.exist?(schedule_file)
+    Sidekiq::Cron::Job.load_from_hash YAML.load_file(schedule_file)
+  end
+end
+
+Sidekiq.configure_client do |config|
+  config.redis = {url: ENV["REDIS_SIDEKIQ_URL"]}
+end