+++ spec/support/contentful_helpers.rb
@@ -6,9 +6,24 @@ module ContentfulHelpers
     raw_response = File.read("#{Rails.root}/spec/fixtures/contentful/#{fixture_filename}")
 
     contentful_connector = stub_contentful_connector
-    contentful_response = fake_contentful_step_entry(contentful_fixture_filename: fixture_filename)
+    contentful_response = fake_contentful_entry(contentful_fixture_filename: fixture_filename)
     allow(contentful_connector).to receive(:get_entry_by_id)
       .with(entry_id)
+      .and_return(contentful_response)
+
+    allow(contentful_response).to receive(:raw)
+      .and_return(raw_response)
+  end
+
+  def stub_get_contentful_entries(
+    entry_id: "5kZ9hIFDvNCEhjWs72SFwj",
+    fixture_filename: "multiple-entries-example.json"
+  )
+    raw_response = File.read("#{Rails.root}/spec/fixtures/contentful/#{fixture_filename}")
+
+    contentful_connector = stub_contentful_connector
+    contentful_response = fake_contentful_entry_array(contentful_fixture_filename: fixture_filename)
+    allow(contentful_connector).to receive(:get_all_entries)
       .and_return(contentful_response)
 
     allow(contentful_response).to receive(:raw)
@@ -30,13 +45,7 @@ module ContentfulHelpers
     contentful_client
   end
 
-  def stub_contentful_step(fake_entry: fake_contentful_step_entry)
-    get_contentful_step_double = instance_double(GetContentfulEntry)
-    allow(GetContentfulEntry).to receive(:new).and_return(get_contentful_step_double)
-    allow(get_contentful_step_double).to receive(:call).and_return(fake_entry)
-  end
-
-  def fake_contentful_step_entry(contentful_fixture_filename:)
+  def fake_contentful_entry(contentful_fixture_filename:)
     raw_response = File.read("#{Rails.root}/spec/fixtures/contentful/#{contentful_fixture_filename}")
     hash_response = JSON.parse(raw_response)
 
@@ -49,8 +58,16 @@ module ContentfulHelpers
       options: hash_response.dig("fields", "options"),
       type: hash_response.dig("fields", "type"),
       next: double(id: hash_response.dig("fields", "next", "sys", "id")),
-      raw: raw_response,
+      primary_call_to_action: hash_response.dig("fields", "primaryCallToAction"),
+      raw: hash_response,
       content_type: double(id: hash_response.dig("sys", "contentType", "sys", "id"))
     )
+  end
+
+  def fake_contentful_entry_array(contentful_fixture_filename:)
+    raw_response = File.read("#{Rails.root}/spec/fixtures/contentful/#{contentful_fixture_filename}")
+    response_hash = JSON.parse(raw_response)
+
+    Contentful::ResourceBuilder.new(response_hash).run
   end
 end